Add `money` module
* move `#toCents` to `money.js` module from `stripe.js`
* modify: `#toCents` returns 0 for non-dollar-value strings
* create `#toDollarStr` method to convert from cents to formatted str
